+++ internal/server/exposed/handlers/version.go
